Make the RESET history action special.
Previously, the RESET action was encoded in the same way as MOVE and
WALL. To do so, RESET needed to be allocated space in the control data
(which types the action), increasing the control data from 1 bit (just
MOVE and WALL) to 2 bits (MOVE, WALL, RESET).

Making RESET a special case allows the control data to be 1 bit, which,
with another optimization, will allow the WALL action to be encoded in 1
byte instead of 2.